# perf(otelstorage): optimize allocations in `Hash`