--- a/internal/otelstorage/hash.go
+++ b/internal/otelstorage/hash.go
@@ -25,13 +25,6 @@ func AttrHash(m pcommon.Map) Hash {
 	return h.Sum128().Bytes()
 }
 
-// StrHash computes string hash.
-func StrHash(s string) Hash {
-	h := xxh3.New()
-	_, _ = h.WriteString(s)
-	return h.Sum128().Bytes()
-}
-
 func hashValue(h *xxh3.Hasher, val pcommon.Value) {
 	var buf [8]byte
 
@@ -76,7 +69,12 @@ func hashMap(h *xxh3.Hasher, m pcommon.Map) {
 		key   string
 		value pcommon.Value
 	}
-	pairs := make([]pair, 0, m.Len())
+	var pairs []pair
+	if l := m.Len(); l < 16 {
+		pairs = make([]pair, 0, 16)
+	} else {
+		pairs = make([]pair, 0, m.Len())
+	}
 	m.Range(func(k string, v pcommon.Value) bool {
 		pairs = append(pairs, pair{
 			key:   k,