For #293.

```
                 │   old.txt   │               new.txt               │
                 │   sec/op    │   sec/op     vs base                │
_metricsBatch-32   1.257m ± 0%   1.095m ± 1%  -12.83% (p=0.000 n=15)

                 │   old.txt    │               new.txt                │
                 │     B/op     │     B/op      vs base                │
_metricsBatch-32   511.2Ki ± 0%   290.8Ki ± 0%  -43.11% (p=0.000 n=15)

                 │   old.txt   │               new.txt               │
                 │  allocs/op  │  allocs/op   vs base                │
_metricsBatch-32   7.894k ± 0%   6.280k ± 0%  -20.45% (p=0.000 n=15)
```